# Conflicts:
#	src/main/java/com/igomall/controller/admin/FileController.java
#	src/main/java/com/igomall/controller/admin/course/CourseController.java
#	src/main/java/com/igomall/controller/common/course/CourseController.java
#	src/main/java/com/igomall/controller/member/IndexController.java
#	src/main/java/com/igomall/entity/course/Course.java
#	src/main/java/com/igomall/entity/course/Lesson.java
#	src/main/java/com/igomall/service/FileService.java
#	src/main/java/com/igomall/service/impl/FileServiceImpl.java
#	src/main/java/com/igomall/service/impl/SearchServiceImpl.java
#	src/main/java/com/igomall/util/SystemUtils.java

+++ src/main/java/com/igomall/Demo08.java
@@ -7,24 +7,26 @@ import java.io.File;
 public class Demo08 {
     public static void main(String[] args) throws Exception{
         String path = "/Users/blackboy/Desktop/0411/中国刑事大案纪实_刑事案件要案记录";
-
-
         File parent = new File(path);
-
         File[] files = parent.listFiles();
         for (File file :files) {
             String fileName = file.getName();
-            //String[] names = fileName.split(" _ ");
-            //if(names.length==3){
-                String newFileName = fileName.replaceAll("-","_");
+            fileName = fileName.replaceAll(" ","");
+            String[] names = fileName.split("-");
+            System.out.println(names.length);
+            if(names.length==3){
+                fileName = names[1]+".mp3";
+                System.out.println(fileName);
                 try{
-                    File file1 = new File(path,newFileName);
+                    File file1 = new File(path,fileName);
                     FileUtils.copyFile(file,file1);
                     FileUtils.deleteQuietly(file);
+                    System.out.println("=================================");
                 }catch (Exception e){
                     e.printStackTrace();
                 }
-            //}
+            }
+
         }
 
     }
